projeto1/requirements.txt
ipython==9.1.0
openai==1.97.0
pandas==2.3.1
pdf2image==1.17.0
python-dotenv==1.1.1
